# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtefactOut": {
        "description": "The `artefact` read model \u2014 the synthesised evidence base.\n\nArgs:\n    title: Artefact title.\n    question: The evidence question the artefact answers.\n    coverage_snapshot: Embedded coverage snapshot.\n    sections: Artefact sections, in final page order.\n    references: Numbered reference list.",
        "properties": {
          "coverage_snapshot": {
            "$ref": "#/components/schemas/CoverageSnapshotOut"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "references": {
            "items": {
              "$ref": "#/components/schemas/ReferenceOut"
            },
            "title": "References",
            "type": "array"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SectionOut"
            },
            "title": "Sections",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "question",
          "coverage_snapshot"
        ],
        "title": "ArtefactOut",
        "type": "object"
      },
...
      "ChunkContextOut": {
        "description": "The chunk-context read model for a cited span (the 008 seam).\n\nArgs:\n    context: Context text, clamped to a character window around the\n        cited span.\n    span_start: Start offset of the clamped context.\n    span_end: End offset of the clamped context.\n    clamped: Whether the window was clamped (hit a chunk boundary).",
        "properties": {
          "clamped": {
            "title": "Clamped",
            "type": "boolean"
          },
          "context": {
            "title": "Context",
            "type": "string"
          },
          "span_end": {
            "title": "Span End",
            "type": "integer"
          },
          "span_start": {
            "title": "Span Start",
            "type": "integer"
          }
        },
        "required": [
          "context",
          "span_start",
          "span_end",
          "clamped"
        ],
        "title": "ChunkContextOut",
        "type": "object"
      },
...
      "CoverageOut": {
        "description": "The `coverage` read model \u2014 the composed one-line coverage sentence.\n\nArgs:\n    sentence: The composed coverage sentence (stop condition + adequacy).\n    base: Structured basis the sentence was composed from.",
        "properties": {
          "base": {
            "additionalProperties": true,
            "title": "Base",
            "type": "object"
          },
          "sentence": {
            "title": "Sentence",
            "type": "string"
          }
        },
        "required": [
          "sentence"
        ],
        "title": "CoverageOut",
        "type": "object"
      },
...
      "FunnelOut": {
        "description": "The acquisition-to-citation funnel counts.\n\nArgs:\n    found: Total acquired.\n    relevant: Passed abstract screening.\n    screened_out: Excluded at screening.\n    quality_checked: Passed appraisal.\n    read_in_full: Full text ingested.\n    selected: Passed selection.\n    findings: Findings extracted.\n    cited: Cited in the artefact.\n\nAny field is `None` before its stage has run \u2014 the funnel spans the\nfull flow, unlike the screened-in-only `landscape`.",
        "properties": {
          "cited": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cited"
          },
          "findings": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Findings"
          },
          "found": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Found"
          },
          "quality_checked": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Checked"
          },
          "read_in_full": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read In Full"
          },
          "relevant": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevant"
          },
          "screened_out": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Screened Out"
          },
          "selected": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected"
          }
        },
        "title": "FunnelOut",
        "type": "object"
      },
...
      "GroupsOut": {
        "description": "The `groups` read model.\n\nArgs:\n    facets: Per-facet group listings.",
        "properties": {
          "facets": {
            "items": {
              "$ref": "#/components/schemas/FacetGroupsOut"
            },
            "title": "Facets",
            "type": "array"
          }
        },
        "title": "GroupsOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LandscapeOut": {
        "description": "Distributions over the screened-in set only (never the found count).\n\nArgs:\n    evidence_types: Counts by evidence type label.\n    years: Counts by publication year.\n    themes: Landscape themes.\n    geographies: Optional counts by geography label.",
        "properties": {
          "evidence_types": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Evidence Types",
            "type": "object"
          },
          "geographies": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geographies"
          },
          "themes": {
            "items": {
              "$ref": "#/components/schemas/ThemeOut"
            },
            "title": "Themes",
            "type": "array"
          },
          "years": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Years",
            "type": "object"
          }
        },
        "title": "LandscapeOut",
        "type": "object"
      },
...
      "Page_DecisionOut_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/DecisionOut"
            },
            "title": "Data",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "pagination"
        ],
        "title": "Page[DecisionOut]",
        "type": "object"
      },
      "Page_EvidenceItemOut_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItemOut"
            },
            "title": "Data",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "pagination"
        ],
        "title": "Page[EvidenceItemOut]",
        "type": "object"
      },
      "Page_FindingOut_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "title": "Data",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "pagination"
        ],
        "title": "Page[FindingOut]",
        "type": "object"
      },
      "Page_ProjectOut_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ProjectOut"
            },
            "title": "Data",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "pagination"
        ],
        "title": "Page[ProjectOut]",
        "type": "object"
      },
...
      "ProjectCreate": {
        "additionalProperties": false,
        "description": "Inbound body for `POST /api/v1/projects`.\n\nArgs:\n    name: Project display name, 1-200 characters. Outer whitespace is\n        stripped before the length constraint is applied\n        (`str_strip_whitespace`).\n    question: Optional initial evidence question.",
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProjectCreate",
        "type": "object"
      },
      "ProjectOut": {
        "description": "A project resource.\n\nArgs:\n    project_id: The project's identity.\n    name: Current display name.\n    question: Current evidence question, or `None` if not yet set.\n    status: Lifecycle status (`active` unless archived; no hard delete).\n    created_at: When the project was created.\n    updated_at: When the project row was last written.\n    archived_at: When the project was archived, or `None` if active.\n    latest_run: The derived latest-run read model, or `None` before any\n        run has been created.",
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "latest_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LatestRun"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "status": {
            "enum": [
              "active",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "name",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectOut",
        "type": "object"
      },
...
      "ThemeOut": {
        "description": "One landscape theme.\n\nArgs:\n    name: Theme name.\n    size: Number of items in the theme.\n    description: Short theme description.",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "size",
          "description"
        ],
        "title": "ThemeOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Policy Atlas API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/projects": {
      "get": {
        "description": "List the authenticated user's projects with derived latest-run state.",
        "operationId": "list_projects_api_v1_projects_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "active",
              "enum": [
                "active",
                "archived",
                "all"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProjectOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "description": "Create one active project owned by the authenticated subject.",
        "operationId": "create_project_api_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/artefact": {
      "get": {
        "description": "Return the latest persisted synthesis artefact or a shaped absence.",
        "operationId": "artefact_api_v1_projects__project_id__artefact_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtefactOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Artefact",
        "tags": [
          "read-models"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/citations/{citation_key}/context": {
      "get": {
        "description": "Return a clamped context window for an artefact citation id.",
        "operationId": "chunk_context_api_v1_projects__project_id__citations__citation_key__context_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "citation_key",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Citation Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkContextOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Chunk Context",
        "tags": [
          "read-models"
        ]
      }
    },
    "/api/v1/projects/{project_id}/coverage": {
      "get": {
        "description": "Return the composed latest search coverage statement.",
        "operationId": "coverage_api_v1_projects__project_id__coverage_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Coverage",
        "tags": [
          "read-models"
        ]
      }
    },
    "/api/v1/projects/{project_id}/decisions": {
      "get": {
        "description": "Return the allowlisted audit and steering decision history.",
        "operationId": "decisions_api_v1_projects__project_id__decisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_DecisionOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Decisions",
        "tags": [
          "read-models"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/evidence": {
      "get": {
        "description": "Return a bounded page from the evidence status ladder.",
        "operationId": "evidence_api_v1_projects__project_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_EvidenceItemOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Evidence",
        "tags": [
          "read-models"
        ]
      }
    },
    "/api/v1/projects/{project_id}/findings": {
      "get": {
        "description": "Return a bounded page of IOF and ICF findings.",
        "operationId": "findings_api_v1_projects__project_id__findings_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_FindingOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Findings",
        "tags": [
          "read-models"
        ]
      }
    },
    "/api/v1/projects/{project_id}/funnel": {
      "get": {
        "description": "Return the durable acquisition-to-citation funnel.",
        "operationId": "funnel_api_v1_projects__project_id__funnel_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FunnelOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Funnel",
        "tags": [
          "read-models"
        ]
      }
    },
    "/api/v1/projects/{project_id}/groups": {
      "get": {
        "description": "Return the latest grouping facets and residual counts.",
        "operationId": "groups_api_v1_projects__project_id__groups_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Groups",
        "tags": [
          "read-models"
        ]
      }
    },
    "/api/v1/projects/{project_id}/landscape": {
      "get": {
        "description": "Return screened-in-only landscape distributions.",
        "operationId": "landscape_api_v1_projects__project_id__landscape_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LandscapeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Landscape",
        "tags": [
          "read-models"
        ]
      }
    },
...
  }
}